Add combatants helpers
