+++ combat/engine/__init__.py
@@ -1,6 +1,6 @@
 """Engine subpackage for combat components."""
 
-from .common import _current_hp, CombatParticipant
+from ..combatants import _current_hp, CombatParticipant
 from .turn_manager import TurnManager
 from .aggro_tracker import AggroTracker
 from .damage_processor import DamageProcessor
